requirements.txt
streamlit
torch
torchvision
transformers
pytorch_wavelets
pandas
numpy
matplotlib
scikit-learn
Pillow
